requirements.txt
beautifulsoup4==4.12.2
Django==4.2.3
fake-useragent==1.1.3
lxml==4.9.3
pandas==2.0.3
PyYAML==6.0.1
requests==2.31.0
xmltodict==0.13.0
